moved viewport component to highest level in file structure
