+++ lib/src/index.ts
@@ -10,4 +10,5 @@ export * from './History'
 export * from './Sidebar'
 export * from './Toolbar'
 export * from './utils'
+export * from './Viewport'
 export * from './VisGraph'
