+++ lib/src/layout/index.ts
@@ -3,5 +3,4 @@ export * from './Edge'
 export * from './Point'
 export * from './Rectangle'
 export * from './Vertex'
-export * from './Viewport'
 export * from './tools'
